create display and edit component, seperate concerns more
